Выполнил правки в оформлении, разбил код на файлы
