+++ package.json
@@ -31,6 +31,7 @@
     "not < 0.25%"
   ],
   "dependencies": {
-    "dayjs": "1.11.7"
+    "dayjs": "1.11.7",
+    "nanoid": "4.0.2"
   }
 }
