+++ src/mock/const.js
@@ -1,0 +1,55 @@
+const RoutePointTypes = [
+  'taxi',
+  'bus',
+  'train',
+  'ship',
+  'drive',
+  'flight',
+  'check-in',
+  'sightseeing',
+  'restaurant',
+];
+
+const DestinationNames = [
+  'Abakan',
+  'Barnaul',
+  'Veliky Novgorod',
+  'Gatchina',
+  'Dmitrov',
+  'Yekaterinburg',
+  'Zlatoust',
+  'Izhevsk',
+  'Yoshkar-Ola',
+  'Kazan',
+  'Lipetsk',
+  'Magadan',
+  'Norilsk',
+  'Orenburg',
+  'Pskov',
+  'Rostov-on-Don',
+  'Saint Petersburg',
+  'Tolyatti',
+  'Ulan-Ude',
+  'Khabarovsk',
+  'Cheboksary',
+  'Shuya',
+  'Elista',
+  'Yuzhno-Sakhalinsk',
+  'Yakutsk',
+];
+
+const Descriptions = [
+  'Lorem ipsum dolor sit amet, consectetur adipiscing elit. ',
+  'Cras aliquet varius magna, non porta ligula feugiat eget. ',
+  'Fusce tristique felis at fermentum pharetra. ',
+  'Aliquam id orci ut lectus varius viverra. ',
+  'Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. ',
+  'Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. ',
+  'Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. ',
+  'Sed sed nisi sed augue convallis suscipit in sed felis. ',
+  'Aliquam erat volutpat. ',
+  'Nunc fermentum tortor ac porta dapibus. ',
+  'In rutrum ac purus sit amet tempus. ',
+];
+
+export {RoutePointTypes, DestinationNames, Descriptions};
