+++ src/mock/util.js
@@ -1,0 +1,21 @@
+import dayjs from 'dayjs';
+
+const getRandomInteger = (a = 0, b = 1) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+
+  return Math.floor(lower + Math.random() * (upper - lower + 1));
+};
+
+const generateDate = () => {
+  const maxDaysGap = 6;
+  const maxHoursGap = 1;
+  const maxMinutesGap = 30;
+  const daysGap = getRandomInteger(-maxDaysGap, maxDaysGap);
+  const hoursGap = getRandomInteger(-maxHoursGap, maxHoursGap);
+  const minutesGap = getRandomInteger(-maxMinutesGap, maxMinutesGap);
+
+  return dayjs().add(daysGap, 'day').add(hoursGap, 'hour').add(minutesGap, 'minute');
+};
+
+export { getRandomInteger, generateDate };
