+++ src/model/destinations-model.js
@@ -1,0 +1,9 @@
+import { getDestinations } from '../mock/destination-mock.js';
+
+export default class DestinationsModel {
+  destinations = getDestinations();
+
+  getDestinations() {
+    return this.destinations;
+  }
+}
