+++ src/model/offers-model.js
@@ -1,0 +1,9 @@
+import { getOffers } from '../mock/offer-mock.js';
+
+export default class OffersModel {
+  offers = getOffers();
+
+  getOffers() {
+    return this.offers;
+  }
+}
