+++ src/view/points-list-view.js
@@ -1,12 +1,12 @@
 import { createElement } from '../render.js';
 
-function createEventsListTemplate() {
+function createPointsListTemplate() {
   return '<ul class="trip-events__list"></ul>';
 }
 
-export default class EventsListView {
+export default class PointsListView {
   getTemplate() {
-    return createEventsListTemplate();
+    return createPointsListTemplate();
   }
 
   getElement() {
